+++ voxygen/anim/src/golem/alpha.rs
@@ -7,7 +7,7 @@ use common::states::utils::StageSection;
 pub struct AlphaAnimation;
 
 impl Animation for AlphaAnimation {
-    type Dependency = (Option<StageSection>, f64, f64);
+    type Dependency = (Option<StageSection>, f32, f32);
     type Skeleton = GolemSkeleton;
 
     #[cfg(feature = "use-dyn-lib")]
@@ -18,23 +18,23 @@ impl Animation for AlphaAnimation {
     fn update_skeleton_inner(
         skeleton: &Self::Skeleton,
         (stage_section, global_time, timer): Self::Dependency,
-        anim_time: f64,
+        anim_time: f32,
         _rate: &mut f32,
         _s_a: &SkeletonAttr,
     ) -> Self::Skeleton {
         let mut next = (*skeleton).clone();
 
         let (move1base, move2base, move3) = match stage_section {
-            Some(StageSection::Buildup) => ((anim_time as f32).powf(0.25), 0.0, 0.0),
-            Some(StageSection::Swing) => (1.0, anim_time as f32, 0.0),
-            Some(StageSection::Recover) => (1.0, 1.0, (anim_time as f32).powf(4.0)),
+            Some(StageSection::Buildup) => (anim_time.powf(0.25), 0.0, 0.0),
+            Some(StageSection::Swing) => (1.0, anim_time, 0.0),
+            Some(StageSection::Recover) => (1.0, 1.0, anim_time.powf(4.0)),
             _ => (0.0, 0.0, 0.0),
         };
 
         let pullback = 1.0 - move3;
         let subtract = global_time - timer;
         let check = subtract - subtract.trunc();
-        let mirror = (check - 0.5).signum() as f32;
+        let mirror = (check - 0.5).signum();
 
         let move1 = move1base * pullback;
         let move2 = move2base * pullback;
