+++ voxygen/anim/src/object/shoot.rs
@@ -26,7 +26,7 @@ impl Animation for ShootAnimation {
     fn update_skeleton_inner(
         skeleton: &Self::Skeleton,
         (_active_tool_kind, _second_tool_kind, stage_section, body): Self::Dependency,
-        anim_time: f64,
+        anim_time: f32,
         rate: &mut f32,
         s_a: &SkeletonAttr,
     ) -> Self::Skeleton {
@@ -35,9 +35,9 @@ impl Animation for ShootAnimation {
         let mut next = (*skeleton).clone();
 
         let (movement1, movement2, movement3) = match stage_section {
-            Some(StageSection::Buildup) => (anim_time as f32, 0.0, 0.0),
-            Some(StageSection::Swing) => (1.0, (anim_time as f32).powf(0.25), 0.0),
-            Some(StageSection::Recover) => (1.0, 1.0, anim_time as f32),
+            Some(StageSection::Buildup) => (anim_time, 0.0, 0.0),
+            Some(StageSection::Swing) => (1.0, anim_time.powf(0.25), 0.0),
+            Some(StageSection::Recover) => (1.0, 1.0, anim_time),
             _ => (0.0, 0.0, 0.0),
         };
