+++ voxygen/anim/src/quadruped_low/dash.rs
@@ -9,7 +9,7 @@ use std::f32::consts::PI;
 pub struct DashAnimation;
 
 impl Animation for DashAnimation {
-    type Dependency = (f32, f64, Option<StageSection>, f64);
+    type Dependency = (f32, f32, Option<StageSection>, f32);
     type Skeleton = QuadrupedLowSkeleton;
 
     #[cfg(feature = "use-dyn-lib")]
@@ -19,23 +19,23 @@ impl Animation for DashAnimation {
     fn update_skeleton_inner(
         skeleton: &Self::Skeleton,
         (_velocity, global_time, stage_section, timer): Self::Dependency,
-        anim_time: f64,
+        anim_time: f32,
         _rate: &mut f32,
         _s_a: &SkeletonAttr,
     ) -> Self::Skeleton {
         let mut next = (*skeleton).clone();
 
         let (movement1base, chargemovementbase, movement2base, movement3) = match stage_section {
-            Some(StageSection::Buildup) => ((anim_time as f32).sqrt(), 0.0, 0.0, 0.0),
+            Some(StageSection::Buildup) => (anim_time.sqrt(), 0.0, 0.0, 0.0),
             Some(StageSection::Charge) => (1.0, 1.0, 0.0, 0.0),
-            Some(StageSection::Swing) => (1.0, 1.0, (anim_time as f32).powi(4), 0.0),
-            Some(StageSection::Recover) => (1.0, 1.0, 1.0, anim_time as f32),
+            Some(StageSection::Swing) => (1.0, 1.0, anim_time.powi(4), 0.0),
+            Some(StageSection::Recover) => (1.0, 1.0, 1.0, anim_time),
             _ => (0.0, 0.0, 0.0, 0.0),
         };
         let pullback = 1.0 - movement3;
         let subtract = global_time - timer;
         let check = subtract - subtract.trunc();
-        let mirror = (check - 0.5).signum() as f32;
+        let mirror = (check - 0.5).signum();
         let twitch1 = (mirror * movement1base * 9.5).sin();
         let twitch1fast = (mirror * movement1base * 25.0).sin();
         //let twitch3 = (mirror * movement3 * 4.0).sin();
@@ -43,14 +43,12 @@ impl Animation for DashAnimation {
         //let movement2 = mirror * movement2base * pullback;
         let movement1abs = movement1base * pullback;
         let movement2abs = movement2base * pullback;
-        let short = (((1.0)
-            / (0.72 + 0.28 * ((anim_time as f32 * 16.0_f32 + PI * 0.25).sin()).powi(2)))
-        .sqrt())
-            * ((anim_time as f32 * 16.0_f32 + PI * 0.25).sin())
+        let short = ((1.0 / (0.72 + 0.28 * ((anim_time * 16.0_f32 + PI * 0.25).sin()).powi(2)))
+            .sqrt())
+            * ((anim_time * 16.0_f32 + PI * 0.25).sin())
             * chargemovementbase
             * pullback;
-        let shortalt =
-            (anim_time as f32 * 16.0_f32 + PI * 0.25).sin() * chargemovementbase * pullback;
+        let shortalt = (anim_time * 16.0_f32 + PI * 0.25).sin() * chargemovementbase * pullback;
 
         next.head_upper.orientation =
             Quaternion::rotation_x(movement1abs * 0.4 + movement2abs * 0.3)
