+++ voxygen/anim/src/quadruped_medium/beta.rs
@@ -7,7 +7,7 @@ use common::states::utils::StageSection;
 pub struct BetaAnimation;
 
 impl Animation for BetaAnimation {
-    type Dependency = (f32, f64, Option<StageSection>, f64);
+    type Dependency = (f32, f32, Option<StageSection>, f32);
     type Skeleton = QuadrupedMediumSkeleton;
 
     #[cfg(feature = "use-dyn-lib")]
@@ -17,7 +17,7 @@ impl Animation for BetaAnimation {
     fn update_skeleton_inner(
         skeleton: &Self::Skeleton,
         (velocity, global_time, stage_section, timer): Self::Dependency,
-        anim_time: f64,
+        anim_time: f32,
         _rate: &mut f32,
         s_a: &SkeletonAttr,
     ) -> Self::Skeleton {
@@ -25,15 +25,15 @@ impl Animation for BetaAnimation {
         let speed = (Vec2::<f32>::from(velocity).magnitude()).min(24.0);
 
         let (movement1base, movement2base, movement3) = match stage_section {
-            Some(StageSection::Buildup) => ((anim_time as f32).powf(0.25), 0.0, 0.0),
-            Some(StageSection::Swing) => (1.0, (anim_time as f32).sqrt(), 0.0),
-            Some(StageSection::Recover) => (1.0, 1.0, (anim_time as f32).powi(4)),
+            Some(StageSection::Buildup) => (anim_time.powf(0.25), 0.0, 0.0),
+            Some(StageSection::Swing) => (1.0, anim_time.sqrt(), 0.0),
+            Some(StageSection::Recover) => (1.0, 1.0, anim_time.powi(4)),
             _ => (0.0, 0.0, 0.0),
         };
         let pullback = 1.0 - movement3;
         let subtract = global_time - timer;
         let check = subtract - subtract.trunc();
-        let mirror = (check - 0.5).signum() as f32;
+        let mirror = (check - 0.5).signum();
         let movement1 = movement1base * mirror * pullback;
         let movement1abs = movement1base * pullback;
         let movement2 = movement2base * mirror * pullback;
